fix: isolate WAL open resource cleanup (#51765)
